Add multi-provider football API support (Sportmonks + API-Football)
Introduces a provider abstraction so users can select their data source via
FOOTBALL_API_PROVIDER env var. Adds FootballApiProviderInterface, ApiFootballProvider,
and SportmonksProvider, rewrites SoccerAPIController to delegate to the bound
provider, and adds the Sportmonks SDK back as an optional dependency.
