+++ .env.example
@@ -29,6 +29,12 @@ MAIL_ENCRYPTION=null
 MAIL_FROM_ADDRESS="hello@example.com"
 MAIL_FROM_NAME="${APP_NAME}"
 
+FOOTBALL_API_PROVIDER=api-football
+
 # API-Football — get a free key at https://dashboard.api-football.com/register
 API_FOOTBALL_KEY=
 API_FOOTBALL_TIMEZONE=Europe/Amsterdam
+
+# Sportmonks — get a key at https://sportmonks.com
+# SPORTMONKS_FOOTBALL_API_TOKEN=
+# SPORTMONKS_FOOTBALL_TIMEZONE=UTC
