+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,9 @@
 
 namespace App\Providers;
 
+use App\Contracts\FootballApiProviderInterface;
+use App\Services\FootballApi\ApiFootballProvider;
+use App\Services\FootballApi\SportmonksProvider;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -13,6 +16,11 @@ class AppServiceProvider extends ServiceProvider
 
     public function register(): void
     {
-        //
+        $this->app->bind(FootballApiProviderInterface::class, function () {
+            return match (config('football-api.provider')) {
+                'sportmonks' => new SportmonksProvider(),
+                default => new ApiFootballProvider(),
+            };
+        });
     }
 }
