+++ composer.json
@@ -9,7 +9,8 @@
         "php": "^8.3",
         "jenssegers/agent": "^2.6",
         "laravel/framework": "^12.0",
-        "laravel/tinker": "^3.0.2"
+        "laravel/tinker": "^3.0.2",
+        "pyaesoneaung/sportmonks-football-api": "^3.2"
     },
     "require-dev": {
         "fakerphp/faker": "^1.23",
