--- config/api-football.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'base_url' => 'https://v3.football.api-sports.io/',
-    'api_key' => env('API_FOOTBALL_KEY'),
-    'timezone' => env('API_FOOTBALL_TIMEZONE', 'UTC'),
-];
