+++ config/sportmonks-football-api.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'base_url' => 'https://api.sportmonks.com/v3/',
+    'api_token' => env('SPORTMONKS_FOOTBALL_API_TOKEN'),
+    'timezone' => env('SPORTMONKS_FOOTBALL_TIMEZONE', 'UTC'),
+    'return_type' => 'array',
+];
